Read EXIF metadata to create a simple feature object or write EXIF metadata to image files

read_exif() read EXIF data from folder of files. This function also assigns a cardinal direction based on the direction metadata and recodes the orientation metadata.

For write_exif() the parameters are used to multiple tags with the same values:

title: Title, IPTC:Headline, IPTC:ObjectName, XMP-dc:Title

description: ImageDescription, XMP-dc:Description, and IPTC:Caption-Abstract

keywords: Keywords, IPTC:Keywords, XMP-dc:Subject

Usage

read_exif(
  path = NULL,
  fileext = NULL,
  tags = getOption("read_exif.tags", default = default_tags),
  ...
)

write_exif(
  path,
  fileext = NULL,
  title = NULL,
  author = NULL,
  date = NULL,
  keywords = NULL,
  description = NULL,
  args = NULL,
  overwrite = TRUE,
  append_keywords = FALSE
)

Arguments

path

A path to folder or file.

fileext

The file extension or file type; defaults to NULL.

tags

Optional list of EXIF tags to read from files. Must include GPS tags if to create an sf object based on the resulting data.frame object.

...

Additional EXIF tags to pass to exiftoolr::exif_read()

title

Title to add to file metadata with exiftoolr, Default: NULL.

author

Author to add to file metadata with exiftoolr, Default: NULL.

date

Date to add to file metadata with exiftoolr (not currently working), Default: NULL.

keywords

Keyword(s) added to file metadata with with exiftoolr, Default: NULL.

description

Description added to file metadata.

args

Alternate arguments passed to exiftoolr::exif_call(). If args is not NULL, title, author, date, and keywords are ignored; defaults to NULL.

overwrite

If TRUE, overwrite any existing EXIF metadata present in the provided fields; defaults to TRUE

append_keywords

If TRUE, append keywords, if FALSE, replace keywords in file metadata.